Implemented role of Interface In services
