+++ src/main/java/edutech/backend/repository/UserRepository.java
@@ -2,9 +2,11 @@ package edutech.backend.repository;
 
 import edutech.backend.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
+@Repository
 public interface UserRepository extends JpaRepository<User, Long> {
 
     // Method to check if a user exists by username
